+++ plugin/app/Models/Submission.php
@@ -94,6 +94,19 @@ class Submission extends Model
         return $this->hasMany(TestSuite::class);
     }
 
+    public function unitTests()
+    {
+        return $this->hasManyThrough(UnitTest::class, TestSuite::class,
+            'charon_test_suite.submission_id', 'charon_unit_test.test_suite_id',
+            'id', 'id');
+    }
+
+    public function reviewComments()
+    {
+        return $this->hasManyThrough(ReviewComment::class, SubmissionFile::class,
+            'charon_submission_file.submission_id', 'charon_review_comment.submission_file_id');
+    }
+
     public function getGitTimestampAttribute($gitTimestamp)
     {
         $gitTimestamp = Carbon::parse( $gitTimestamp, 'UTC');