+++ plugin/resources/assets/js/pages/assignment/assignment.js
@@ -15,6 +15,7 @@ import 'vue-good-table/dist/vue-good-table.css'
 import 'vue-material/dist/vue-material.min.css'
 import 'vue-material/dist/theme/default.css'
 import CodeTemplates from "./components/CodeTemplates.vue";
+import ToggleButton from "../../components/partials/ToggleButton.vue";
 
 Vue.directive('highlightjs', HighlightDirective);
 Vue.use(VueMaterial)
@@ -34,7 +35,7 @@ const app = new Vue({
     el: '#app',
     vuetify: new Vuetify(),
 
-    components: {AssignmentView, CodeEditor, CharonTabs, CharonTab, GradesCheckboxes, CodeTemplates},
+    components: {AssignmentView, CodeEditor, CharonTabs, CharonTab, GradesCheckboxes, CodeTemplates, ToggleButton},
 
     icons: {
         iconfont: 'mdi', // 'mdi' || 'mdiSvg' || 'md' || 'fa' || 'fa4'